Dodałem funkcje skalowania i inicjalizowania robota w wybranym przez uzytkownika miejscu. Do zrobinia doxy i porzadek w kodzie, zastanowic sie nad optymalizacja masowego tworzenia obiektow Wektor2D.

+++ Wektor2D.hh
@@ -14,6 +14,6 @@
 /*!
  * \brief Alias do instacji szablon SWektor
  */
-typedef SWektor<float,2>  Wektor2D;
+typedef SWektor<double ,2>  Wektor2D;
 
 #endif
